+++ src/body/CollisionBody.cpp
@@ -137,6 +137,7 @@ uint CollisionBody::getNbColliders() const {
 
 // Return a const pointer to a given collider of the body
 /**
+* @param index Index of a Collider of the body
 * @return The const pointer of a given collider of the body
 */
 const Collider* CollisionBody::getCollider(uint colliderIndex) const {
@@ -150,6 +151,7 @@ const Collider* CollisionBody::getCollider(uint colliderIndex) const {
 
 // Return a pointer to a given collider of the body
 /**
+* @param index Index of a Collider of the body
 * @return The pointer of a given collider of the body
 */
 Collider* CollisionBody::getCollider(uint colliderIndex) {